Close the release gaps found by auditing tessl as a package, not a workspace crate
`cargo package` did not verify. build.rs writes the offline-compatibility
`default.metallib` to the crate root, which is outside OUT_DIR, and Cargo
rejects that while verifying a tarball ("Source directory was modified by
build.rs during cargo publish"). The copy is now skipped when the manifest dir
is a packaging directory, where it is useless anyway: a consumer building from a
registry always compiles the metallib fresh into OUT_DIR, and only this
repository's offline runs set TESSL_SKIP_AOT. `cargo package` now verifies
clean — the crate compiles from its own tarball, kernels included.

`#![allow(dead_code)]` at the crate root was hiding exactly two things, and a
blanket suppression is the wrong answer to both:
  - GpuTensor's `storage`/`runtime` are lifetime anchors the MTLTensor borrows.
    Never read on purpose, so they now say so with a scoped allow.
  - GpuBuffer::kind() had no caller in any of the three crates. BufferKind is
    already public API and the pool a buffer came from changes what its holder
    may do with it on drop, so it is promoted to public rather than deleted.

mtl_tensor is now behind an off-by-default `quant-prep` feature. Its main entry
point returns Err("not wired yet"), which is not something to ship as public API
in 0.1.0. The gate immediately found a consumer an earlier grep had missed —
gemma-metal calls the *re-exported* `tessl::nax_verify_readiness`, not
`tessl::mtl_tensor::*` — so gemma-metal now opts in explicitly. That is the gate
doing its job on its first run.

TESSL_* is now canonical for all 15 runtime env vars, not just the 5 GEMM ones.
Legacy METAL_RUNTIME_*, METAL_NATIVE_* and GEMMA_METAL_* spellings are still
read, in that order, so nothing that works today stops working. The compile-time
metallib variable is renamed TESSL_METALLIB (internal, no compatibility
surface); TESSL_SKIP_AOT still honours METAL_RUNTIME_SKIP_AOT since it is set by
hand.

The two arch_02 parity tests that need golden/fwd/ (~80 MB, gitignored because
regenerable) died on a bare "No such file or directory" naming one .npy. They
now fail with the path, the reason it is absent from a fresh clone, and the
exact command that regenerates it. A check that cannot run has to say what would
make it run.

Docs: 20 rustdoc warnings to 0 — matrix shapes like A[M,K] in doc comments were
being parsed as intra-doc links. Clippy autofix applied. README documents the
feature, the full env var table, and states plainly that the decode/ICB modules
are NOT gated in 0.1.0 and why: they are woven into the binder hot path, and
splitting them is a change to code that must be correct rather than a packaging
tidy-up.

Verified: tessl 65 tests (68 with quant-prep), tessl-arch02 49, gemma-metal 134.
Audit PASS. Fuzz soak 4 seeds x 1200 cases clean. 0 dead-code warnings,
0 rustdoc warnings, `cargo package` verifies.

+++ Rust_MLKit/crates/tessl/src/mtl_tensor.rs
@@ -95,8 +95,14 @@ pub struct GpuTensor {
     pub tensor: Retained<ProtocolObject<dyn MTLTensor>>,
     pub dtype: MTLTensorDataType,
     pub dims: Vec<usize>,
-    /// When buffer-backed, keep the storage alive.
+    // Lifetime anchors, never read: the MTLTensor above borrows this storage,
+    // and the runtime owns the allocator that storage came from. Dropping
+    // either while `tensor` is live is a use-after-free, so they are held, not
+    // used. Scoped `allow` rather than a crate-level one, which would hide the
+    // next genuinely dead field.
+    #[allow(dead_code)]
     pub(crate) storage: Option<GpuBuffer>,
+    #[allow(dead_code)]
     pub(crate) runtime: Arc<GpuRuntime>,
 }
 
@@ -166,7 +172,7 @@ pub unsafe fn tensor_from_buffer(
     let mtl_dtype = dtype.to_mtl()?;
     let desc = make_descriptor(dims, mtl_dtype, MTLTensorUsage::Compute)?;
     let align = rt.device.tensorSizeAndAlignWithDescriptor(&desc);
-    if byte_offset % (align.align as usize) != 0 {
+    if byte_offset % align.align != 0 {
         return Err(format!(
             "tensor buffer offset {byte_offset} not aligned to {}",
             align.align
